requirements.txt
# Core
Django==3.2.25
djangorestframework==3.12.4
channels==3.0.4
daphne==3.0.2
channels-redis==3.3.1

# Authentication & User Management
phonenumbers==8.13.37
django-phonenumber-field==5.2.0

# Security
cryptography==36.0.2
pyopenssl==22.0.0

# Media & File Handling
Pillow==8.4.0

# Utilities
python-dateutil==2.8.2
pytz==2022.1